requirements.txt
asgiref==3.7.2
certifi==2024.2.2
Django==5.0.3
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-nine==0.2.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
elastic-transport==8.12.0
elasticsearch==8.12.1
elasticsearch-dsl==8.12.0
gunicorn==21.2.0
packaging==23.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
urllib3==2.2.1
